requirements.txt
paho-mqtt
python-dotenv
requests
fastapi
uvicorn
